rename db_config dirs
