+++ schematic_db/db_config/__init__.py
@@ -1,0 +1,4 @@
+"""
+g
+"""
+from .db_config import *
