--- schematic_db/db_object_config/__init__.py
@@ -1,2 +1,0 @@
-"""Import all from module"""
-from .db_object_config import *
